requirements.txt
streamlit>=1.37.0
google-genai>=1.0.0
pydantic>=2.8.0